fix: version browser-swarm extension worker
